Moved the launcher into Sharpcraft project.
