+++ Sharpcraft/Launcher.resx
