+++ Sharpcraft/Sharpcraft.csproj
@@ -115,6 +115,12 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Launcher.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Launcher.Designer.cs">
+      <DependentUpon>Launcher.cs</DependentUpon>
+    </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Sharpcraft.cs" />
@@ -132,7 +138,7 @@
       <Project>{6B60AF16-B510-4D54-BC50-CC232B6826AE}</Project>
       <Name>Sharpcraft.Networking</Name>
     </ProjectReference>
-    <ProjectReference Include="..\Sharpcraft.SteamGUI\Sharpcraft.Forms.csproj">
+    <ProjectReference Include="..\Sharpcraft.Forms\Sharpcraft.Forms.csproj">
       <Project>{5B58612A-455C-4A60-8B2C-196F274E7EBA}</Project>
       <Name>Sharpcraft.Forms</Name>
     </ProjectReference>
@@ -173,11 +179,16 @@
       <Install>true</Install>
     </BootstrapperPackage>
   </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="Launcher.resx">
+      <DependentUpon>Launcher.cs</DependentUpon>
+    </EmbeddedResource>
+  </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <Import Project="$(MSBuildExtensionsPath)\Microsoft\XNA Game Studio\Microsoft.Xna.GameStudio.targets" />
   <ProjectExtensions>
     <VisualStudio>
-      <UserProperties BuildVersion_BuildVersioningStyle="None.None.Increment.TimeStamp" BuildVersion_UpdateAssemblyVersion="True" BuildVersion_UpdateFileVersion="True" />
+      <UserProperties BuildVersion_UpdateFileVersion="True" BuildVersion_UpdateAssemblyVersion="True" BuildVersion_BuildVersioningStyle="None.None.Increment.TimeStamp" />
     </VisualStudio>
   </ProjectExtensions>
   <!--
